pyproject.toml
[project]
name = "config-loader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]
